✨ Refactoring of content's data structure to support multiple courses, years and semesters.
