+++ src/Interfaces/Resource.ts
