+++ src/Interfaces/ResourceLinkType.ts
