+++ src/Interfaces/Semester.ts
@@ -1,0 +1,6 @@
+import { Subject } from "./Subject"
+
+export interface Semester {
+	name: string
+	subjects: Subject[] | undefined
+}
